+++ src/settings/TagSettings.js
@@ -11,6 +11,12 @@ import {
 import { ConfigManager } from '@folio/stripes/smart-components';
 import { useStripes } from '@folio/stripes/core';
 
+import { useConfigurationSave } from '../hooks/useConfigurationSave';
+import {
+  TAGS_SCOPE,
+  TAGS_KEY,
+} from '../constants';
+
 const propTypes = {
   label: PropTypes.string,
 };
@@ -19,18 +25,28 @@ const TagSettings = ({ label }) => {
   const stripes = useStripes();
   const intl = useIntl();
   const ConfigManagerConnected = useMemo(() => stripes.connect(ConfigManager), [stripes]);
+  const { saveConfiguration, setAction } = useConfigurationSave();
 
   const getInitialValues = (settings) => {
-    const value = !settings.length || settings[0].value === 'true';
+    setAction(settings[0]?.id);
+
+    const value = !settings.length || settings[0].value;
     return { tags_enabled: value };
+  };
+
+  // Save the flag in configurations API as well for backward compatibility.
+  // This will be removed in future releases once all UI modules have migrated to settings API (FOLIO-4343).
+  const handleAfterSave = (setting) => {
+    saveConfiguration(setting);
   };
 
   return (
     <ConfigManagerConnected
       getInitialValues={getInitialValues}
       label={label}
-      moduleName="TAGS"
-      configName="tags_enabled"
+      scope={TAGS_SCOPE}
+      configName={TAGS_KEY}
+      onAfterSave={handleAfterSave}
     >
       <Row>
         <Col xs={12}>