+++ test/jest/__mock__/stripesCore.mock.js
@@ -112,7 +112,7 @@ jest.mock('@folio/stripes/core', () => {
     AppContextMenu,
     useOkapiKy,
     useNamespace,
-    useStripes: () => STRIPES,
+    useStripes: jest.fn().mockReturnValue(STRIPES),
   };
 }, { virtual: true });