+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change history for ui-tags
 
+## IN PROGRESS
+* Save the `tags_enabled` flag in both mod-settings and mod-configuration to allow apps to also migrate to mod-settings and not break those that haven't migrated yet. Refs UITAG-96.
+
 ## [9.0.0](https://github.com/folio-org/ui-tags/tree/v9.0.0) (2025-03-14)
 [Full Changelog](https://github.com/folio-org/ui-tags/compare/v8.2.1...v9.0.0)
 * *BREAKING* Migrate stripes dependencies to their Sunflower versions. Refs UITAG-86.